Removed the JSSImporter dependency
Removing the JSSImporter dependency (finally)!
OfflineApps will now use a new base Processor, MountShare, for mounting shares.
I apologize to those using this...  To provide the _ability_ to support more than just SMB shares, I made some changes to numerous things to support that if it was ever needed.
This includes a breaking change for those using the mounting option as I updated input variable names to remove SMB from the names, but this shouldn't be _too_ big of a deal, hopefully.