Shenanigans with the download URL
Elixxier's website defaults to a Hong Kong based CDN and then has some JavaScript that replaces the URL after the page loads....which breaks this for AutoPkg's web scraping methods...

This fixes it for me and allows you to update the download URL to your locale if desired.  You'll have to check Elixxier's website to determine this.