Change installation method...
This recipe will now expect AutoCAD to be installed using the official installation method.  See the recipe description for a link to the install script that can be used to install AutoCAD.

The previous installation method (that worked for v2021 and v2022) stopped working with v2023.  It was really a hack work around in the first palce.

The only problem with this method is, it does not work when it is executed from a .pkg's postinstall script for some odd reason.  So the script will need to be ran separately from the Install App's distribution method.